requirements_gradio.txt
agno==1.5.10
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.10
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
attrs==25.3.0
bandit==1.8.3
blinker==1.9.0
boltons==21.0.0
boolean.py==5.0
bracex==2.5.post1
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-option-group==0.5.7
colorama==0.4.6
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
Deprecated==1.2.18
detect-secrets==1.5.0
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
frozenlist==1.6.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
glom==22.1.0
googleapis-common-protos==1.70.0
gradio==5.33.0
gradio_client==1.10.2
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==7.1.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
license-expression==30.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.9.0
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.4
narwhals==1.41.1
numpy==2.2.6
openai==1.84.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.18
packageurl-python==0.17.1
packaging==24.2
pandas==2.3.0
pbr==6.1.1
peewee==3.18.1
pillow==11.2.1
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
propcache==0.3.1
protobuf==4.25.8
psutil==7.0.0
py-serializable==2.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rich==13.5.3
rpds-py==0.25.1
ruamel.yaml==0.18.13
ruamel.yaml.clib==0.2.12
ruff==0.11.13
safehttpx==0.1.6
semantic-version==2.10.0
semgrep==1.124.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==2.3.6
starlette==0.46.2
stevedore==5.4.1
streamlit==1.45.1
tenacity==9.1.2
toml==0.10.2
tomli==2.0.2
tomlkit==0.13.3
tornado==6.5.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
wcmatch==8.5.2
websockets==15.0.1
wrapt==1.17.2
yarl==1.20.0
zipp==3.22.0
